# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "CustomerController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cari nama/telepon/email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["INDIVIDUAL", "BUSINESS"]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["ACTIVE", "INACTIVE", "MERGED"]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "displayName",
              "type": "string",
              "enum": ["displayName", "createdAt", "updatedAt"]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Daftar/cari pelanggan",
        "tags": ["customers"]
      },
      "post": {
        "operationId": "CustomerController_create",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buat pelanggan (telepon/email dinormalisasi otomatis)",
        "tags": ["customers"]
      }
    },
    "/customers/merge-candidates": {
      "get": {
        "operationId": "CustomerController_candidates",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Daftar kandidat duplikat (default: PENDING)",
        "tags": ["customers"]
      }
    },
    "/customers/merge-history": {
      "get": {
        "operationId": "CustomerController_history",
        "parameters": [
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Riwayat merge pelanggan",
        "tags": ["customers"]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomerController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail pelanggan (identitas, alamat, indikator duplikat)",
        "tags": ["customers"]
      },
      "patch": {
        "operationId": "CustomerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ubah pelanggan",
        "tags": ["customers"]
      },
      "delete": {
        "operationId": "CustomerController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Arsipkan pelanggan (soft delete)",
        "tags": ["customers"]
      }
    },
    "/customers/{id}/identities": {
      "post": {
        "operationId": "CustomerController_addIdentity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIdentityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah identitas kanal pelanggan",
        "tags": ["customers"]
      }
    },
    "/customers/identities/{identityId}": {
      "patch": {
        "operationId": "CustomerController_updateIdentity",
        "parameters": [
          {
            "name": "identityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIdentityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ubah status verifikasi / primary identitas",
        "tags": ["customers"]
      }
    },
    "/customers/{id}/addresses": {
      "post": {
        "operationId": "CustomerController_addAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah alamat pelanggan",
        "tags": ["customers"]
      }
    },
    "/customers/merge-candidates/{candidateId}": {
      "patch": {
        "operationId": "CustomerController_reviewCandidate",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review kandidat duplikat (konfirmasi/tolak/abaikan)",
        "tags": ["customers"]
      }
    },
    "/customers/merge/preview": {
      "post": {
        "operationId": "CustomerController_mergePreview",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergePreviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview hasil merge dua pelanggan (tanpa mengubah data)",
        "tags": ["customers"]
      }
    },
    "/customers/merge/execute": {
      "post": {
        "operationId": "CustomerController_mergeExecute",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeExecuteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eksekusi merge (atomic, dengan history + audit)",
        "tags": ["customers"]
      }
    },
...
  },
  "info": {
    "title": "FlowNiaga API",
    "description": "API FlowNiaga — Sistem Operasi Bisnis Omnichannel. Endpoint tenant-scoped membutuhkan header x-tenant-id.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["INDIVIDUAL", "BUSINESS"],
            "default": "INDIVIDUAL"
          },
          "displayName": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "primaryPhone": {
            "type": "string",
            "example": "081234567890"
          },
          "primaryEmail": {
            "type": "string",
            "example": "budi@contoh.id"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "consentStatus": {
            "type": "string",
            "enum": ["UNKNOWN", "OPTED_IN", "OPTED_OUT"]
          },
          "preferredLanguage": {
            "type": "string",
            "default": "id"
          },
          "preferredChannel": {
            "type": "string",
            "example": "whatsapp"
          }
        },
        "required": ["displayName"]
      },
      "PatchCustomerDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["INDIVIDUAL", "BUSINESS"]
          },
          "displayName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "primaryPhone": {
            "type": "string"
          },
          "primaryEmail": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "consentStatus": {
            "type": "string",
            "enum": ["UNKNOWN", "OPTED_IN", "OPTED_OUT"]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE"]
          },
          "preferredLanguage": {
            "type": "string"
          },
          "preferredChannel": {
            "type": "string"
          }
        }
      },
      "AddIdentityDto": {
        "type": "object",
        "properties": {
          "identityType": {
            "type": "string",
            "enum": ["PHONE", "EMAIL", "MARKETPLACE_ACCOUNT", "WHATSAPP", "MANUAL_REFERENCE"]
          },
          "value": {
            "type": "string",
            "example": "0812-3456-7890",
            "description": "Nilai apa adanya; dinormalisasi server"
          },
          "channelId": {
            "type": "string",
            "description": "Kanal asal (mis. mock marketplace)"
          },
          "externalId": {
            "type": "string",
            "description": "ID akun eksternal pada kanal"
          },
          "isPrimary": {
            "type": "boolean",
            "default": false
          },
          "verificationStatus": {
            "type": "string",
            "enum": ["UNVERIFIED", "VERIFIED"],
            "default": "UNVERIFIED"
          }
        },
        "required": ["identityType", "value"]
      },
      "UpdateIdentityDto": {
        "type": "object",
        "properties": {}
      },
      "AddAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "default": "Utama"
          },
          "recipientName": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "phone": {
            "type": "string"
          },
          "addressLine": {
            "type": "string",
            "example": "Jl. Melati No. 10 RT 02/RW 05"
          },
          "village": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "example": "Jakarta Pusat"
          },
          "province": {
            "type": "string",
            "example": "DKI Jakarta"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "default": "ID"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "isPrimary": {
            "type": "boolean",
            "default": false
          }
        },
        "required": ["recipientName", "addressLine", "city", "province"]
      },
      "ReviewCandidateDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["CONFIRMED_DUPLICATE", "REJECTED", "IGNORED"]
          }
        },
        "required": ["status"]
      },
      "MergePreviewDto": {
        "type": "object",
        "properties": {
          "targetCustomerId": {
            "type": "string",
            "description": "Customer yang dipertahankan (master)"
          },
          "sourceCustomerId": {
            "type": "string",
            "description": "Customer yang digabungkan (source, akan ditandai MERGED)"
          },
          "keepFromSource": {
            "description": "Field yang diambil dari SOURCE (selain itu nilai target dipertahankan). Contoh: [\"primaryPhone\",\"notes\"]",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["targetCustomerId", "sourceCustomerId"]
      },
      "MergeExecuteDto": {
        "type": "object",
        "properties": {
          "targetCustomerId": {
            "type": "string",
            "description": "Customer yang dipertahankan (master)"
          },
          "sourceCustomerId": {
            "type": "string",
            "description": "Customer yang digabungkan (source, akan ditandai MERGED)"
          },
          "keepFromSource": {
            "description": "Field yang diambil dari SOURCE (selain itu nilai target dipertahankan). Contoh: [\"primaryPhone\",\"notes\"]",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string",
            "example": "Nomor telepon sama; dikonfirmasi pelanggan yang sama."
          }
        },
        "required": ["targetCustomerId", "sourceCustomerId", "reason"]
      }
    }
  }
}